[Fix/#7] SearchView 코드 분리 및 수정 등
